replace dao pattern
