+++ juntagrico_assignment_request/admins/assignment_request_admin.py
@@ -5,8 +5,8 @@ from juntagrico.admins import BaseAdmin
 from juntagrico.util.admin import MyHTMLWidget
 from juntagrico.config import Config
 
-from juntagrico_assignment_request.dao.assignmentrequestdao import AssignmentRequestDao
 from juntagrico_assignment_request.entity.assignment_request import AssignmentRequest
+from juntagrico_assignment_request.utils import all_approvers
 
 
 class AssignmentRequestAdminForm(ModelForm):
@@ -16,7 +16,7 @@ class AssignmentRequestAdminForm(ModelForm):
 
     def __init__(self, *a, **k):
         super().__init__(*a, **k)
-        self.fields['approver'].queryset = AssignmentRequestDao.all_approvers()
+        self.fields['approver'].queryset = all_approvers()
         instance = k.get('instance')
         self.fields['assignment_link'].initial = self.get_assignment_link(instance)
