+++ juntagrico_assignment_request/mailer/__init__.py
@@ -1,6 +1,6 @@
 from juntagrico.config import Config
 
-from juntagrico_assignment_request.dao.assignmentrequestdao import AssignmentRequestDao
+from juntagrico_assignment_request.utils import all_notified_on_unapproved_assignments
 
 '''
 Server generated Emails
@@ -14,7 +14,7 @@ def get_approver_emails(assignment_request):
         receivers.append(assignment_request.approver.email)
     else:
         # Otherwise inform all that should be notified on unapproved assignments
-        for notified in AssignmentRequestDao.all_notified_on_unapproved_assignments():
+        for notified in all_notified_on_unapproved_assignments():
             receivers.append(notified.email)
     if not receivers:
         # Nobody? Send it to the info mail then
